# faster internal version of update_tibble_attrs()