--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -64,7 +64,9 @@ VignetteBuilder:
 RdMacros: 
     lifecycle
 Remotes: 
-    r-lib/pillar
+    r-lib/pillar, 
+    r-lib/vctrs,		
+    r-lib/rlang
 Encoding: UTF-8
 LazyData: yes
 Roxygen: list(markdown = TRUE)