--- a/R/subsetting.R
+++ b/R/subsetting.R
@@ -589,20 +589,6 @@ vectbl_recycle_rows <- function(x, n, j, name) {
   abort(error_inconsistent_cols(n, name, size, "Existing data"))
 }
 
-attrs_names_only <- c("names", "row.names")
-
 vectbl_restore <- function(xo, x) {
-  n <- fast_nrow(xo)
-
-  # FIXME: Use new_tibble()?
-  forbidden <- attrs_names_only
-  attrs <- attributes(x)
-  attrs <- attrs[!(names(attrs) %in% forbidden)]
-
-  attributes(xo)[names(attrs)] <- attrs
-
-  # Need to patch manually
-  attr(xo, "row.names") <- .set_row_names(n)
-
-  xo
+  .Call(`tibble_restore_impl`, xo, x)
 }