--- a/src/attributes.c
+++ b/src/attributes.c
@@ -1,6 +1,8 @@
 #include "tibble.h"
 
 SEXP tibble_update_attrs(SEXP x, SEXP dots) {
+  x = PROTECT(Rf_shallow_duplicate(x));
+
   while(!Rf_isNull(dots)) {
     SEXP tag = TAG(dots);
     if (!Rf_isNull(tag)) {
@@ -8,5 +10,6 @@ SEXP tibble_update_attrs(SEXP x, SEXP dots) {
     }
     dots = CDR(dots);
   }
+  UNPROTECT(1);
   return x;
 }