--- a/src/attributes.c
+++ b/src/attributes.c
@@ -27,17 +27,6 @@ SEXP tibble_restore_impl(SEXP xo, SEXP x) {
     attr_x = CDR(attr_x);
   }
 
-  // set row.names if needed
-  SEXP rn = Rf_getAttrib(xo, R_RowNamesSymbol);
-  if(!(Rf_isInteger(rn) && LENGTH(rn) == 2 && INTEGER(rn)[0] == NA_INTEGER) && rn != R_NilValue) {
-    int n = LENGTH(rn);
-    rn = PROTECT(Rf_allocVector(INTSXP, 2));
-    INTEGER(rn)[0] = NA_INTEGER;
-    INTEGER(rn)[1] = -n;
-    Rf_setAttrib(xo, R_RowNamesSymbol, rn);
-    UNPROTECT(1);
-  }
-
   UNPROTECT(1);
   return xo;
 }