--- a/src/tibble.h
+++ b/src/tibble.h
@@ -6,5 +6,6 @@
 
 SEXP tibble_matrixToDataFrame(SEXP xSEXP);
 SEXP tibble_string_to_indices(SEXP x);
+SEXP tibble_update_attrs(SEXP x, SEXP dots);
 
 #endif /* TIBBLE_H */