--- a/src/tibble.h
+++ b/src/tibble.h
@@ -7,5 +7,6 @@
 SEXP tibble_matrixToDataFrame(SEXP xSEXP);
 SEXP tibble_string_to_indices(SEXP x);
 SEXP tibble_update_attrs(SEXP x, SEXP dots);
+SEXP tibble_restore_impl(SEXP xo, SEXP x);
 
 #endif /* TIBBLE_H */