This has come up as part of performance benchmarks for dplyr 1.0.0: 

``` r
library(tibble)
library(rlang)

update_tibble_attrs_old <- function(x, ...) {
  # Compat for https://github.com/r-spatial/sf/pull/1232
  attribs <- list(...)
  if (has_length(attribs) && is_named(attribs)) {
    attributes(x)[names(attribs)] <- attribs
  }
  
  x
}
update_tibble_attrs <- tibble:::update_tibble_attrs


df <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100)))
groups <- tibble(g = unique(df$g))

bench::mark(
  new = update_tibble_attrs(df, groups = groups), 
  old = update_tibble_attrs_old(df, groups = groups) 
)
#> # A tibble: 2 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new          2.09µs   2.72µs   344885.        0B      0  
#> 2 old          3.98ms   4.96ms      201.    3.81MB     18.7
```

<sup>Created on 2020-02-13 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>

I believe this is because the form `attributes(x)[names(attribs)] <- ` pays for setting *all* the attributes, including `row.names` which is expensive to get/set. 

closes tidyverse/dplyr#4855