@krlmlr 2020-02-14
Thanks! I have removed compat code that is no longer necessary with the sf update. It's still very slow, though.

I have also added a test regarding overwriting of attributes. I think our implementation will need to account for this. Do we need to adapt?