@romainfrancois 2020-02-14
`structure()` seems to be as slow: 

``` r
library(tibble)
library(rlang)

update_tibble_attrs_old <- function(x, ...) {
  # Compat for https://github.com/r-spatial/sf/pull/1232
  attribs <- list(...)
  if (has_length(attribs) && is_named(attribs)) {
    attributes(x)[names(attribs)] <- attribs
  }
  x
}
update_tibble_attrs <- tibble:::update_tibble_attrs

df <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100)))
groups <- tibble(g = unique(df$g))

bench::mark(
  new = update_tibble_attrs(df, groups = groups),
  old = update_tibble_attrs_old(df, groups = groups), 
  str = structure(df, groups = groups)
)
#> # A tibble: 3 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 new          2.02µs   2.71µs   335090.        0B      0  
#> 2 old          3.97ms   4.77ms      211.    3.82MB     22.4
#> 3 str          3.96ms    4.3ms      226.    3.81MB     15.8
```

<sup>Created on 2020-02-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>