@romainfrancois 2020-02-14
I believe `vectbl_restore()` has the same issue, i.e. using `attributes()[<-`

```r
vectbl_restore <- function(xo, x) {
  n <- fast_nrow(xo)

  # FIXME: Use new_tibble()?
  forbidden <- attrs_names_only
  attrs <- attributes(x)
  attrs <- attrs[!(names(attrs) %in% forbidden)]

  attributes(xo)[names(attrs)] <- attrs

  # Need to patch manually
  attr(xo, "row.names") <- .set_row_names(n)

  xo
}
```