@krlmlr 2020-02-14
Good catch, pushed the tests.

I still wonder why `vectbl_restore()` seems to be much faster:

``` r
library(tibble)

x <- tibble(a = 1:3)
y <- x

vectbl_restore <- tibble:::vectbl_restore

bench::mark(vectbl_restore(x, y), y)
#> # A tibble: 2 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vectbl_restore(x, y)    5.8µs   6.93µs   123302.    3.97KB     24.7
#> 2 y                      23.1ns  28.17ns 23492926.        0B      0
```

<sup>Created on 2020-02-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>