@romainfrancois 2020-02-14
For similar reasons, I've updated `vec_tbl_restore()` : 

``` r
library(tibble)
library(rlang)

df1 <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100)))
df2 <- tibble(x = rnorm(1e6))

vectbl_restore <- tibble:::vectbl_restore
fast_nrow <- function(x) {
  .row_names_info(x, 2L)
}

attrs_names_only <- c("names", "row.names")
vectbl_restore_old <- function(xo, x) {
  n <- fast_nrow(xo)
  
  # FIXME: Use new_tibble()?
  forbidden <- attrs_names_only
  attrs <- attributes(x)
  attrs <- attrs[!(names(attrs) %in% forbidden)]
  
  attributes(xo)[names(attrs)] <- attrs
  
  # Need to patch manually
  attr(xo, "row.names") <- .set_row_names(n)
  
  xo
}

bench::mark(
  vectbl_restore(df1, df2), 
  vectbl_restore_old(df1, df2)
)
#> # A tibble: 2 x 6
#>   expression                        min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vectbl_restore(df1, df2)        929ns 1.27µs   753454.        0B      0  
#> 2 vectbl_restore_old(df1, df2)   3.97ms 4.81ms      209.    3.81MB     22.7
```

<sup>Created on 2020-02-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>