@romainfrancois 2020-02-14
closes https://github.com/tidyverse/dplyr/issues/4855