@krlmlr 2020-02-14
Oh, it's only a problem with long data frames:

``` r
library(tibble)
library(rlang)

df1 <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100)))
df1a <- df1[1:10000, ]
df1b <- df1[1:1000, ]
df1c <- df1[1:100, ]

vectbl_restore <- tibble:::vectbl_restore

bench::mark(
  vectbl_restore(df1, df1),
  vectbl_restore(df1a, df1a),
  vectbl_restore(df1b, df1b),
  vectbl_restore(df1c, df1c),
  check = FALSE
)
#> # A tibble: 4 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vectbl_restore(df1, df1)     3.57ms   3.97ms      241.    3.81MB     23.4
#> 2 vectbl_restore(df1a, df1a)  40.44µs  43.49µs    20181.   39.11KB     16.4
#> 3 vectbl_restore(df1b, df1b)   9.54µs  11.04µs    75970.    3.95KB     15.2
#> 4 vectbl_restore(df1c, df1c)   6.44µs   7.39µs   118001.      448B     23.6
```

<sup>Created on 2020-02-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>