@romainfrancois 2020-02-14
yes @krlmlr I suppose this is `row.names` being inconveninentnly materialized when using `attributes[]<-`