@romainfrancois 2020-02-14
The errors are unrelated: 

```
Last 13 lines of output:
83
  x[125]: "Error: Must extract column with a single subscript."
84
  y[125]: "Error: Must extract column with a single valid subscript."
85
  
86
  x[129]: "Error: Must extract column with a single subscript."
87
  y[129]: "Error: Must extract column with a single valid subscript."
88
  
89
  x[133]: "Error: Must extract column with a single subscript."
90
  y[133]: "Error: Must extract column with a single valid subscript."
```

I suppose I'm using a younger `vctrs` or `rlang`