@krlmlr 2020-02-14
If you revert 752bfa2 the CI errors should go away.

Applying `[` to a tibble speeds up subsequent execution of `vectbl_restore()`:

``` r
library(tibble)
library(rlang)

df1 <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100)))
df1a <- df1[1:100000, ]
df1b <- df1[1:10000, ]
df1c <- df1[1:1000, ]

vectbl_restore <- tibble:::vectbl_restore

bench::mark(
  vectbl_restore(df1, df1),
  vectbl_restore(df1a, df1a),
  vectbl_restore(df1b, df1b),
  vectbl_restore(df1c, df1c),
  check = FALSE
)
#> # A tibble: 4 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vectbl_restore(df1, df1)     3.54ms   4.05ms      232.    3.81MB     23.4
#> 2 vectbl_restore(df1a, df1a) 345.57µs 401.14µs     2235.  390.67KB     18.3
#> 3 vectbl_restore(df1b, df1b)  40.69µs  46.54µs    19021.   39.11KB     13.2
#> 4 vectbl_restore(df1c, df1c)   9.79µs  11.77µs    71793.    3.95KB     14.4
```

<sup>Created on 2020-02-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0)</sup>

I guess this no longer holds with the C implementation?