@romainfrancois 2020-02-14
``` r
library(tibble)
library(rlang)

df1 <- tibble(x = rnorm(1e6), g = sample(rep(1:1e4, 100)))
df1a <- df1[1:100000, ]
df1b <- df1[1:10000, ]
df1c <- df1[1:1000, ]

vectbl_restore <- tibble:::vectbl_restore

bench::mark(
  vectbl_restore(df1, df1),
  vectbl_restore(df1a, df1a),
  vectbl_restore(df1b, df1b),
  vectbl_restore(df1c, df1c),
  check = FALSE
)
#> # A tibble: 4 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vectbl_restore(df1, df1)     1.12µs   1.38µs   503585.        0B      0  
#> 2 vectbl_restore(df1a, df1a)    896ns    1.2µs   576690.        0B      0  
#> 3 vectbl_restore(df1b, df1b)    895ns   1.09µs   653004.        0B     65.3
#> 4 vectbl_restore(df1c, df1c)   1.15µs   1.45µs   607377.        0B      0
```

<sup>Created on 2020-02-14 by the [reprex package](https://reprex.tidyverse.org) (v0.3.0.9000)</sup>